closer to a solution
