+++ src/huggingface_hub/inference/_client.py
@@ -33,26 +33,17 @@
 # - Images are parsed as PIL.Image for easier manipulation.
 # - Provides a "recommended model" for each task => suboptimal but user-wise quicker to get a first script running.
 # - Only the main parameters are publicly exposed. Power users can always read the docs for more options.
-import base64
-import io
-import json
 import logging
 import time
 import warnings
-from contextlib import contextmanager
 from dataclasses import asdict
-from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
-    BinaryIO,
-    ContextManager,
     Dict,
-    Generator,
     Iterable,
     List,
     Optional,
-    Set,
     Union,
     overload,
 )
@@ -60,16 +51,28 @@ from typing import (
 from requests import HTTPError
 from requests.structures import CaseInsensitiveDict
 
-from ..constants import ENDPOINT, INFERENCE_ENDPOINT
+from ..constants import INFERENCE_ENDPOINT
 from ..utils import (
     BadRequestError,
     build_hf_headers,
     get_session,
     hf_raise_for_status,
-    is_numpy_available,
-    is_pillow_available,
 )
 from ..utils._typing import Literal
+from ._common import (
+    ContentT,
+    InferenceTimeoutError,
+    _b64_encode,
+    _b64_to_image,
+    _bytes_to_dict,
+    _bytes_to_image,
+    _get_recommended_model,
+    _import_numpy,
+    _is_tgi_server,
+    _open_as_binary,
+    _set_as_non_tgi,
+    _stream_text_generation_response,
+)
 from ._text_generation import (
     TextGenerationParameters,
     TextGenerationRequest,
@@ -85,18 +88,6 @@ if TYPE_CHECKING:
     from PIL import Image
 
 logger = logging.getLogger(__name__)
-
-UrlT = str
-PathT = Union[str, Path]
-BinaryT = Union[bytes, BinaryIO]
-ContentT = Union[BinaryT, PathT, UrlT]
-
-# Will be globally fetched only once (see '_fetch_recommended_models')
-_RECOMMENDED_MODELS: Optional[Dict[str, Optional[str]]] = None
-
-
-class InferenceTimeoutError(HTTPError, TimeoutError):
-    """Error raised when a model is unavailable or the request times out."""
 
 
 class InferenceClient:
@@ -1217,173 +1208,3 @@ class InferenceClient:
             # Otherwise, we use the default endpoint
             else f"{INFERENCE_ENDPOINT}/models/{model}"
         )
-
-
-def _get_recommended_model(task: str) -> str:
-    model = _fetch_recommended_models().get(task)
-    if model is None:
-        raise ValueError(
-            f"Task {task} has no recommended task. Please specify a model explicitly. Visit"
-            " https://huggingface.co/tasks for more info."
-        )
-    logger.info(
-        f"Using recommended model {model} for task {task}. Note that it is encouraged to explicitly set"
-        f" `model='{model}'` as the recommended models list might get updated without prior notice."
-    )
-    return model
-
-
-def _fetch_recommended_models() -> Dict[str, Optional[str]]:
-    global _RECOMMENDED_MODELS
-    if _RECOMMENDED_MODELS is None:
-        response = get_session().get(f"{ENDPOINT}/api/tasks", headers=build_hf_headers())
-        hf_raise_for_status(response)
-        _RECOMMENDED_MODELS = {
-            task: _first_or_none(details["widgetModels"]) for task, details in response.json().items()
-        }
-    return _RECOMMENDED_MODELS
-
-
-@overload
-def _open_as_binary(content: ContentT) -> ContextManager[BinaryT]:
-    ...  # means "if input is not None, output is not None"
-
-
-@overload
-def _open_as_binary(content: Literal[None]) -> ContextManager[Literal[None]]:
-    ...  # means "if input is None, output is None"
-
-
-@contextmanager  # type: ignore
-def _open_as_binary(content: Optional[ContentT]) -> Generator[Optional[BinaryT], None, None]:
-    """Open `content` as a binary file, either from a URL, a local path, or raw bytes.
-
-    Do nothing if `content` is None,
-
-    TODO: handle a PIL.Image as input
-    TODO: handle base64 as input
-    """
-    # If content is a string => must be either a URL or a path
-    if isinstance(content, str):
-        if content.startswith("https://") or content.startswith("http://"):
-            logger.debug(f"Downloading content from {content}")
-            yield get_session().get(content).content  # TODO: retrieve as stream and pipe to post request ?
-            return
-        content = Path(content)
-        if not content.exists():
-            raise FileNotFoundError(
-                f"File not found at {content}. If `data` is a string, it must either be a URL or a path to a local"
-                " file. To pass raw content, please encode it as bytes first."
-            )
-
-    # If content is a Path => open it
-    if isinstance(content, Path):
-        logger.debug(f"Opening content from {content}")
-        with content.open("rb") as f:
-            yield f
-    else:
-        # Otherwise: already a file-like object or None
-        yield content
-
-
-def _b64_encode(content: ContentT) -> str:
-    """Encode a raw file (image, audio) into base64. Can be byes, an opened file, a path or a URL."""
-    with _open_as_binary(content) as data:
-        data_as_bytes = data if isinstance(data, bytes) else data.read()
-        return base64.b64encode(data_as_bytes).decode()
-
-
-def _b64_to_image(encoded_image: str) -> "Image":
-    """Parse a base64-encoded string into a PIL Image."""
-    Image = _import_pil_image()
-    return Image.open(io.BytesIO(base64.b64decode(encoded_image)))
-
-
-def _bytes_to_dict(content: bytes) -> "Image":
-    """Parse bytes from a Response object into a Python dictionary.
-
-    Expects the response body to be encoded-JSON data.
-    """
-    return json.loads(content.decode())
-
-
-def _bytes_to_image(content: bytes) -> "Image":
-    """Parse bytes from a Response object into a PIL Image.
-
-    Expects the response body to be raw bytes. To deal with b64 encoded images, use `_b64_to_image` instead.
-    """
-    Image = _import_pil_image()
-    return Image.open(io.BytesIO(content))
-
-
-def _stream_text_generation_response(
-    bytes_output_as_lines: Iterable[bytes], details: bool
-) -> Union[Iterable[str], Iterable[TextGenerationStreamResponse]]:
-    # Parse ServerSentEvents
-    for byte_payload in bytes_output_as_lines:
-        # Skip line
-        if byte_payload == b"\n":
-            continue
-
-        payload = byte_payload.decode("utf-8")
-
-        # Event data
-        if payload.startswith("data:"):
-            # Decode payload
-            json_payload = json.loads(payload.lstrip("data:").rstrip("/n"))
-            # Parse payload
-            output = TextGenerationStreamResponse(**json_payload)
-            yield output.token.text if not details else output
-
-
-def _import_pil_image():
-    """Make sure `PIL` is installed on the machine."""
-    if not is_pillow_available():
-        raise ImportError(
-            "Please install Pillow to use deal with images (`pip install Pillow`). If you don't want the image to be"
-            " post-processed, use `client.post(...)` and get the raw response from the server."
-        )
-    from PIL import Image
-
-    return Image
-
-
-def _import_numpy():
-    """Make sure `numpy` is installed on the machine."""
-    if not is_numpy_available():
-        raise ImportError("Please install numpy to use deal with embeddings (`pip install numpy`).")
-    import numpy
-
-    return numpy
-
-
-def _first_or_none(items: List[Any]) -> Optional[Any]:
-    try:
-        return items[0] or None
-    except IndexError:
-        return None
-
-
-# "TGI servers" are servers running with the `text-generation-inference` backend.
-# This backend is the go-to solution to run large language models at scale. However,
-# for some smaller models (e.g. "gpt2") the default `transformers` + `api-inference`
-# solution is still in use.
-#
-# Both approaches have very similar APIs, but not exactly the same. What we do first in
-# the `text_generation` method is to assume the model is served via TGI. If we realize
-# it's not the case (i.e. we receive an HTTP 400 Bad Request), we fallback to the
-# default API with a warning message. We remember for each model if it's a TGI server
-# or not using `_NON_TGI_SERVERS` global variable.
-#
-# For more details, see https://github.com/huggingface/text-generation-inference and
-# https://huggingface.co/docs/api-inference/detailed_parameters#text-generation-task.
-
-_NON_TGI_SERVERS: Set[Optional[str]] = set()
-
-
-def _set_as_non_tgi(model: Optional[str]) -> None:
-    _NON_TGI_SERVERS.add(model)
-
-
-def _is_tgi_server(model: Optional[str]) -> bool:
-    return model not in _NON_TGI_SERVERS
